requirements.txt
tensorflow==2.11.*
tensorflow_text==2.11.*
flask
python-dotenv==1.0.0
numpy<=1.23
flask-cors